changes for refactored CGSim
