+++ CMakeLists.txt
@@ -67,7 +67,7 @@ install(TARGETS CGSim
 )
 
 # Install header files
-install(FILES include/job.h include/DispatcherPlugin.h include/host_extensions.h include/logger.h
+install(FILES include/logger.h include/job.h include/DispatcherPlugin.h include/host_extensions.h
         DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/CGSim)
 
 # Install the executable
