+++ include/PluginLoader.h
@@ -5,7 +5,6 @@
 #include <dlfcn.h>
 #include <libgen.h>
 #include <string>
-#include "logger.h"
 
 template <class Plugin> class PluginLoader
 {
@@ -61,7 +60,6 @@ std::unique_ptr<Plugin> PluginLoader<Plugin>::load(const std::string & pString) 
     typedef void *(*CreationMethod) ();
     auto F = (CreationMethod) f;
     auto * factory = (Plugin *) F();
-    LOG_INFO("Created the dispatcher");
     return std::unique_ptr<Plugin>(factory);
 }
