+++ util/job_executor.cpp
@@ -1,52 +1,26 @@
 #include "job_executor.h"
-#include "logger.h"  
-#include <chrono>
 
-// Recursive function to print a NetZone and its children.
-void printNetZone(const simgrid::s4u::NetZone* zone, int indent = 0) {
-  if (!zone) return;
-
-  std::string indentStr(indent, ' ');
-  LOG_DEBUG("{}Zone Name: {}", indentStr, zone->get_name());
-
-  const std::vector<simgrid::s4u::NetZone*>& children = zone->get_children();
-  if (!children.empty()) {
-    LOG_DEBUG("{}Children:", indentStr);
-    for (const auto child : children) {
-      printNetZone(child, indent + 2);
-    }
-  } else {
-    LOG_DEBUG("{}No children.", indentStr);
-  }
-}
-
-std::unique_ptr<DispatcherPlugin> JOB_EXECUTOR::dispatcher;
-std::unique_ptr<sqliteSaver> JOB_EXECUTOR::saver = std::make_unique<sqliteSaver>();
-// std::vector<Job*> JOB_EXECUTOR::pending_jobs;
-sg4::ActivitySet JOB_EXECUTOR::pending_activities;
-sg4::ActivitySet JOB_EXECUTOR::exec_activities;
+std::unique_ptr<DispatcherPlugin>   JOB_EXECUTOR::dispatcher;
+sg4::ActivitySet                    JOB_EXECUTOR::pending_activities;
+sg4::ActivitySet                    JOB_EXECUTOR::exec_activities;
 
 
-void JOB_EXECUTOR::set_dispatcher(const std::string& dispatcherPath, sg4::NetZone* platform)
+void JOB_EXECUTOR::set_dispatcher(std::unique_ptr<DispatcherPlugin>& d)
 {
-  PluginLoader<DispatcherPlugin> plugin_loader;
-  dispatcher = plugin_loader.load(dispatcherPath);
-
-  // printNetZone(platform);
-  dispatcher->getResourceInformation(platform);
+ dispatcher = std::move(d);
 }
 
 void JOB_EXECUTOR::set_output(const std::string& outputFile)
 {
-  LOG_INFO("Output path set to: {}", outputFile);
-  saver->setFilePath(outputFile);
-  saver->createJobsTable();
+  //LOG_INFO("Output path set to: {}", outputFile);
+  //saver->setFilePath(outputFile);
+  //saver->createJobsTable();
 }
 
-void JOB_EXECUTOR::start_job_execution(JobQueue jobs)
+void JOB_EXECUTOR::start_job_execution(long num_of_jobs_to_run)
 {
   attach_callbacks();
-  LOG_INFO("Callbacks attached. Starting job execution...");
+  //LOG_INFO("Callbacks attached. Starting job execution...");
   const auto* e = sg4::Engine::get_instance();
   sg4::Host* job_server = nullptr;
   for (auto& h : e->get_all_hosts()) {
@@ -57,11 +31,12 @@ void JOB_EXECUTOR::start_job_execution(JobQueue jobs)
   }
 
   if (!job_server) {
-    LOG_CRITICAL("Job Server not initialized properly.");
+    //LOG_CRITICAL("Job Server not initialized properly.");
     exit(-1);
   }
 
-  sg4::Actor::create("JOB-EXECUTOR-actor", job_server, this->start_server, jobs);
+  JobQueue jobs = dispatcher->getWorkload(num_of_jobs_to_run);
+  sg4::Actor::create("JOB-EXECUTOR-actor", job_server, JOB_EXECUTOR::start_server,jobs);
    
   e->run();
 }
@@ -71,7 +46,7 @@ void JOB_EXECUTOR::start_server(JobQueue jobs)
     const auto* e = sg4::Engine::get_instance();
     //sg4::ActivitySet job_activities;
 
-    LOG_INFO("Server started. Initial jobs count: {}", jobs.size());
+    //LOG_INFO("Server started. Initial jobs count: {}", jobs.size());
 
 
     // Transfer all jobs from the queue into a vector for central polling.
@@ -79,12 +54,12 @@ void JOB_EXECUTOR::start_server(JobQueue jobs)
     while (!jobs.empty()) {
         Job* job = jobs.top();
         jobs.pop();
-        saver->saveJob(job);
+        //saver->saveJob(job);
         // LOG_INFO("Job saved to DB: {}", job->id);
 
         // Attempt a one‑time assignment.
         Job* result = dispatcher->assignJob(job);
-        saver->updateJob(result);
+        //saver->updateJob(result);
         if (result->status == "assigned") {
             // If assigned immediately, dispatch it.
             auto fs = simgrid::fsmod::FileSystem::get_file_systems_by_netzone(
@@ -93,7 +68,7 @@ void JOB_EXECUTOR::start_server(JobQueue jobs)
             update_disk_content(fs, job->input_files, job);
             sg4::MessageQueue* mqueue = sg4::MessageQueue::by_name(job->comp_host + "-MQ");
             pending_activities.push(mqueue->put_async(job)->set_name("Comm_Job_" + job->id + "_on_" + job->comp_host));
-            LOG_DEBUG("Job {} dispatched immediately to host {}", job->id, job->comp_host);
+            //LOG_DEBUG("Job {} dispatched immediately to host {}", job->id, job->comp_host);
         } else {
             // Set status and add to pending list.
             if(result->status != "failed"){job->status = "pending"; pending_jobs.push_back(job);}
@@ -112,7 +87,7 @@ void JOB_EXECUTOR::start_server(JobQueue jobs)
         for (auto it = pending_jobs.begin(); it != pending_jobs.end(); ) {
             Job* job = *it;
             dispatcher->assignJob(job);
-            saver->updateJob(job);
+            //saver->updateJob(job);
             retry_counts[job]++;
             if (job->status == "assigned") {
                 auto fs = simgrid::fsmod::FileSystem::get_file_systems_by_netzone(
@@ -121,7 +96,7 @@ void JOB_EXECUTOR::start_server(JobQueue jobs)
                 update_disk_content(fs, job->input_files, job);
                 sg4::MessageQueue* mqueue = sg4::MessageQueue::by_name(job->comp_host + "-MQ");
                 pending_activities.push(mqueue->put_async(job)->set_name("Comm_Job_" + job->id + "_on_" + job->comp_host));
-                LOG_DEBUG("Job {} dispatched after {} retries to host {}", job->id, retry_counts[job], job->comp_host);
+                //LOG_DEBUG("Job {} dispatched after {} retries to host {}", job->id, retry_counts[job], job->comp_host);
                 it = pending_jobs.erase(it);
             }
 
@@ -131,10 +106,10 @@ void JOB_EXECUTOR::start_server(JobQueue jobs)
         }
         if (!exec_activities.empty()) {
           auto activityPtr = exec_activities.wait_any();
-          LOG_INFO("Updated Pending Activities count: {}", JOB_EXECUTOR::pending_activities.size());
-          LOG_INFO("Activity completed: {}", activityPtr->get_name());
+          //LOG_INFO("Updated Pending Activities count: {}", JOB_EXECUTOR::pending_activities.size());
+          //LOG_INFO("Activity completed: {}", activityPtr->get_name());
         }
-        LOG_INFO("Pending jobs count: {}", pending_jobs.size());
+        //LOG_INFO("Pending jobs count: {}", pending_jobs.size());
     }
 
 
@@ -156,24 +131,24 @@ void JOB_EXECUTOR::start_server(JobQueue jobs)
 
   //exec_activities.wait_all();
   pending_activities.wait_all();
-  LOG_DEBUG("Finished All Pending Activities");
+  //LOG_DEBUG("Finished All Pending Activities");
 
 }
 
 void JOB_EXECUTOR::execute_job(Job* j)
 {
-  LOG_DEBUG("Executing job: {}", j->id);
+  //LOG_DEBUG("Executing job: {}", j->id);
   j->status = "running";
-  saver->updateJob(j);
+  //saver->updateJob(j);
 
   const auto* e = sg4::Engine::get_instance();
   auto fs = simgrid::fsmod::FileSystem::get_file_systems_by_netzone(
     e->netzone_by_name_or_null(j->comp_site)).at(j->comp_site + j->comp_host + j->disk + "filesystem");
   // sg4::this_actor::get_host()->extension<HostExtensions>()->registerJob(j);  
   Actions::read_file_async(fs, j, pending_activities, dispatcher);
-  Actions::exec_task_multi_thread_async(j, pending_activities, exec_activities, saver, dispatcher);
+  Actions::exec_task_multi_thread_async(j, pending_activities, exec_activities, dispatcher);
   Actions::write_file_async(fs, j, pending_activities, dispatcher);
-  LOG_DEBUG("Activities added for job: {}", j->job_name);
+  //LOG_DEBUG("Activities added for job: {}", j->job_name);
 
 }
 
@@ -187,7 +162,7 @@ void JOB_EXECUTOR::receiver(const std::string& MQ_name)
     sg4::MessPtr mess = mqueue->get_async();
     mess->wait();
     auto* job = static_cast<Job*>(mess->get_payload());
-    LOG_DEBUG("Received job on queue {}: {}", MQ_name, job->id);
+    //LOG_DEBUG("Received job on queue {}: {}", MQ_name, job->id);
     execute_job(job);
     }
 }
@@ -199,8 +174,8 @@ void JOB_EXECUTOR::start_receivers()
   auto         hosts  = eng->get_all_hosts();
 
   auto host_fetch_time = std::chrono::high_resolution_clock::now();
-  LOG_INFO("Time to fetch hosts: {} ms",
-           std::chrono::duration_cast<std::chrono::milliseconds>(host_fetch_time - start).count());
+  //LOG_INFO("Time to fetch hosts: {} ms",
+       //    std::chrono::duration_cast<std::chrono::milliseconds>(host_fetch_time - start).count());
 
   for (const auto& host : hosts) {
     if (host->get_name() == "JOB-SERVER_cpu-0") continue;
@@ -208,10 +183,10 @@ void JOB_EXECUTOR::start_receivers()
   }
 
   auto end = std::chrono::high_resolution_clock::now();
-  LOG_INFO("Finished creating receivers in: {} ms",
-           std::chrono::duration_cast<std::chrono::milliseconds>(end - host_fetch_time).count());
-  LOG_INFO("Total receiver setup time: {} ms",
-           std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count());
+  //LOG_INFO("Finished creating receivers in: {} ms",
+        //   std::chrono::duration_cast<std::chrono::milliseconds>(end - host_fetch_time).count());
+  //LOG_INFO("Total receiver setup time: {} ms",
+         //  std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count());
 }
 
 void JOB_EXECUTOR::update_disk_content(const std::shared_ptr<simgrid::fsmod::FileSystem>& fs, const std::unordered_map<std::string, size_t>& input_files, Job* j)
@@ -232,25 +207,25 @@ void JOB_EXECUTOR::update_disk_content(const std::shared_ptr<simgrid::fsmod::Fil
   }
 }
 
-void JOB_EXECUTOR::saveJobs(JobQueue jobs)
+/*void JOB_EXECUTOR::saveJobs()
 {
   while (!jobs.empty()) {
     Job* j = jobs.top();
-    saver->updateJob(j);
+    //saver->updateJob(j);
     jobs.pop();
     delete j;
   }
-}
+}*/
 
 void JOB_EXECUTOR::attach_callbacks()
 {
   sg4::Engine::on_simulation_start_cb([]() {
-    LOG_INFO("Simulation starting...");
+    //LOG_INFO("Simulation starting...");
   });
 
   sg4::Engine::on_simulation_end_cb([]() {
-    LOG_INFO("Simulation finished, SIMULATED TIME: {}", sg4::Engine::get_clock());
+    //LOG_INFO("Simulation finished, SIMULATED TIME: {}", sg4::Engine::get_clock());
     dispatcher->onSimulationEnd();
-    saver->exportJobsToCSV();
+    //saver->exportJobsToCSV();
   });
 }
