+++ util/parser.cpp
@@ -1,5 +1,4 @@
 #include "parser.h"
-#include "logger.h"          
 #include <fstream>
 #include <list>
 #include <random>
@@ -69,14 +68,14 @@ void Parser::setSiteNames()
 {
     std::ifstream in(siteInfoFile);
     if (!in.is_open()) {
-        LOG_ERROR("Error: Could not open file {}", siteConnInfoFile);
+        //LOG_ERROR("Error: Could not open file {}", siteConnInfoFile);
         return;
     }
     json j;
     try {
         j = json::parse(in);
     } catch (const json::parse_error& e) {
-        LOG_ERROR("Error parsing JSON: {}", e.what());
+        //LOG_ERROR("Error parsing JSON: {}", e.what());
         return;
     }
     // Get all site names
@@ -85,7 +84,7 @@ void Parser::setSiteNames()
         if (siteCPUCount[site] > 0) {
             site_names.insert(site);
         } else {
-            LOG_INFO("Site Name (invalid or no CPUs or info not available): {}", site);
+            //LOG_INFO("Site Name (invalid or no CPUs or info not available): {}", site);
         }
     }
 }
@@ -94,7 +93,7 @@ void Parser::setSiteNames(const std::list<std::string>& filteredSiteList)
 {
     std::ifstream in(siteInfoFile);
     if (!in.is_open()) {
-        LOG_ERROR("Error: Could not open file {}", siteInfoFile);
+        //LOG_ERROR("Error: Could not open file {}", siteInfoFile);
         return;
     }
     
@@ -102,7 +101,7 @@ void Parser::setSiteNames(const std::list<std::string>& filteredSiteList)
     try {
         j = json::parse(in);
     } catch (const json::parse_error& e) {
-        LOG_ERROR("Error parsing JSON: {}", e.what());
+        //LOG_ERROR("Error parsing JSON: {}", e.what());
         return;
     }
     
@@ -113,7 +112,7 @@ void Parser::setSiteNames(const std::list<std::string>& filteredSiteList)
             if (siteCPUCount[site] > 0) {
                 site_names.insert(site);
             } else {
-                LOG_INFO("Site Name (invalid or no CPUs): {}", site);
+                //LOG_INFO("Site Name (invalid or no CPUs): {}", site);
             }
         }
     } else {
@@ -122,7 +121,7 @@ void Parser::setSiteNames(const std::list<std::string>& filteredSiteList)
             if (siteCPUCount[site] > 0) {
                 site_names.insert(site);
             } else {
-                LOG_INFO("Site Name (invalid or no CPUs): {}", site);
+                //LOG_INFO("Site Name (invalid or no CPUs): {}", site);
             }
         }
     }
@@ -208,7 +207,7 @@ std::unordered_map<std::string, std::unordered_map<std::string, CPUInfo>> Parser
         int num_of_cpus = siteCPUCount[site]; // Little random offset added.
         std::vector<DiskInfo> disks = this->getDisksInfo(site, num_of_cpus);
 
-        LOG_INFO("Adding Site: {} with CPUs: {}", site, num_of_cpus);
+        //LOG_INFO("Adding Site: {} with CPUs: {}", site, num_of_cpus);
         std::vector<double> cpuSpeeds = siteCPUSpeeds[site];
         // std::vector<double> cpuSpeeds = {
         //     1.6e6, 200000, 600000, 1.5e6, 1.4e6, 600000, 2e6, 1.8e6, 1e6, 900000,
@@ -254,7 +253,7 @@ std::unordered_map<std::string, std::unordered_map<std::string, CPUInfo>> Parser
         int num_of_cpus = siteCPUCount[site] + 3; // Little random offset added.
         std::vector<DiskInfo> disks = this->getDisksInfo(site, num_of_cpus);
 
-        LOG_INFO("Adding Site: {} with CPUs: {}", site, num_of_cpus);
+        //LOG_INFO("Adding Site: {} with CPUs: {}", site, num_of_cpus);
         
         // std::cout << "Site CPUS: " << siteCPUCount[site] << std::endl;
         // std::cout << "CPUS count: " << num_of_cpus << std::endl;
@@ -310,104 +309,4 @@ std::unordered_map<std::string, std::pair<double, double>> Parser::getSiteConnIn
     }
  
     return siteConnInfo;
-}
-
-std::vector<std::string> parseCSVLine(const std::string& line) {
-    std::vector<std::string> row;
-    std::string cell;
-    bool in_quotes = false;
-    for (size_t i = 0; i < line.size(); ++i) {
-        char c = line[i];
-        if (c == '"') {
-            in_quotes = !in_quotes; 
-        } else if (c == ',' && !in_quotes) {
-            row.push_back(cell);
-            cell.clear();
-        } else {
-            cell += c;
-        }
-    }
-    row.push_back(cell);
-    for (auto& field : row) {
-        if (!field.empty() && field.front() == '"' && field.back() == '"') {
-            field = field.substr(1, field.size() - 2); 
-        }
-        field.erase(std::remove_if(field.begin(), field.end(),
-                            [](unsigned char c) { return !std::isprint(c); }),
-                            field.end());
-    }
-    return row;
-}
-
-std::priority_queue<Job*> Parser::getJobs(long max_jobs) {
-    std::priority_queue<Job*> jobs;
-    std::ifstream file(jobFile);
-    if (!file.is_open()) {
-        throw std::runtime_error("Could not open file: " + jobFile);
-    }
-    std::string line;
-    std::map<std::string, int> column_map;
-    bool header_parsed = false;
-    while (std::getline(file, line)) {
-        std::vector<std::string> row = parseCSVLine(line);
-        if (max_jobs != -1 && static_cast<long>(jobs.size()) >= max_jobs) {
-            break;
-        }
-        if (!header_parsed) {
-            header_parsed = true;
-            int column_index = 0;
-            for (std::string& column_name : row) {
-                std::transform(column_name.begin(), column_name.end(), column_name.begin(), ::toupper);
-                column_map.insert({column_name, column_index});
-                column_index++;
-            }
-            continue;
-        }
-        try {
-            
-            Job* job = new Job();  // Allocate memory dynamically
-            job->jobid = std::stol(row[column_map.at("PANDAID")]);
-            job->creation_time = row[column_map.at("CREATIONTIME")];
-            job->job_status = row[column_map.at("JOBSTATUS")];
-            job->job_name = row[column_map.at("JOBNAME")];
-            job->cpu_consumption_time = std::stod(row[column_map.at("CPUCONSUMPTIONTIME")]);
-            job->comp_site = row[column_map.at("COMPUTINGSITE")];
-            job->destination_dataset_name = row[column_map.at("DESTINATIONDBLOCK")];
-            job->destination_SE = row[column_map.at("DESTINATIONSE")];
-            job->source_site = row[column_map.at("SOURCESITE")];
-            job->transfer_type = row[column_map.at("TRANSFERTYPE")];
-            job->cores = row[column_map.at("CORECOUNT")].empty() ? 0 : std::stol(row[column_map.at("CORECOUNT")]);
-            job->no_of_inp_files = std::stoi(row[column_map.at("NINPUTDATAFILES")]);
-            job->inp_file_bytes = std::stod(row[column_map.at("INPUTFILEBYTES")]);
-            job->no_of_out_files = std::stoi(row[column_map.at("NOUTPUTDATAFILES")]);
-            job->out_file_bytes = std::stod(row[column_map.at("OUTPUTFILEBYTES")]);
-            job->pilot_error_code = row[column_map.at("PILOTERRORCODE")];
-            job->exe_error_code = row[column_map.at("EXEERRORCODE")];
-            job->ddm_error_code = row[column_map.at("DDMERRORCODE")];
-            job->dispatcher_error_code = row[column_map.at("JOBDISPATCHERERRORCODE")];
-            job->taskbuffer_error_code = row[column_map.at("TASKBUFFERERRORCODE")];
-            job->status = row[column_map.at("JOBSTATUS")];
-          
-            std::string prefix = "/input/user.input." + std::to_string(job->jobid) + ".00000";
-            std::string suffix = ".root";
-            size_t size_per_inp_file = job->inp_file_bytes / job->no_of_inp_files;
-            for (int file = 1; file <= job->no_of_inp_files; file++) {
-                std::string name = prefix + std::to_string(file) + suffix;
-                job->input_files[name] = size_per_inp_file;
-            }
-            prefix = "/output/user.output." + std::to_string(job->jobid) + ".00000";
-            suffix = ".root";
-            size_t size_per_out_file = job->out_file_bytes / job->no_of_out_files;
-            for (int file = 1; file <= job->no_of_out_files; file++) {
-                std::string name = prefix + std::to_string(file) + suffix;
-                job->output_files[name] = size_per_out_file;
-            }
-            jobs.push(job);  // Push pointer to queue
-        } catch (const std::exception& e) {
-            LOG_WARN("Skipping invalid row: {}", line); // currently reasons for being invalid 1) no_of_inp_files,no_of_inp_files  is empty 2)  no_of_out_files,out_file_bytes is empty
-            LOG_WARN("Reason: {}", e.what());
-        }
-    }
-    file.close();
-    return jobs;
 }
